long num added

+++ 0x06-pointers_arrays_strings/101-print_number.c
@@ -7,7 +7,7 @@
 
 void print_number(int n)
 {
-	long num = n;
+	unsigned int num = n;
 
 	if (n < 0)
 	{
